Add Current singleton class and set Current.user
This is a special rails 5+ signleton object, we set the current user to it
after the user gets authenticated in the request.
This way we have access to the current user in our models.
This kind of breaks the seperation of concerns principle of mvc.
But! It inherits from ActiveSupport::CurrentAttributes and is very tread safe.
It is containd and we only use it lightly.

We will use this in the Issue model under the private method.

current_user!

with a bang just in case!!