BIG REFACTOR of the V1::IssuesController
Refactored and removed all the complications, made it very easy to read,
easy to maintain and extend in the future.

Not using a single `if … else` statement not even in the business logic.
Authorisation is now done trough a policy with pundit!
Controller is now simplified and slim, business logic is now clear and in the model where it belongs.

Not a single unnecessary duplication!
Control flow is no longer controlled by Exceptions.
Code is not verbose and almost pure.
Even RuboCop has notting to say, Im only using recommendations from the Ruby/Rails style guides.

I really focused on readability and maintainability.